+++ crates/wdl-lsp/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added `exceptions` to `ServerOptions` ([#542](https://github.com/stjude-rust-labs/wdl/pull/542)).
+
 ## 0.10.0 - 07-09-2025
 
 #### Added